requirements.txt
google-generativeai
Pillow
pytesseract
PyMuPDF

# For pytesseract to function, Tesseract-OCR must be installed on the system.
# You may also need to configure the path to the Tesseract executable.
# For example, in your Python script:
# import pytesseract
# pytesseract.pytesseract.tesseract_cmd = r'/usr/bin/tesseract'  # Or your Tesseract path
# See pytesseract documentation for more details on installation and configuration.
